feat: Move IntegrationRepository to new folder structure
