+++ src/script/components/MessagesList/Message/DeleteMessage.tsx
@@ -19,6 +19,7 @@
 
 import React from 'react';
 
+import {ServiceEntity} from 'Repositories/integration/ServiceEntity';
 import {User} from 'src/script/entity/User';
 import {t} from 'Util/LocalizerUtil';
 import {formatTimeShort, fromUnixTime, TIME_IN_MILLIS} from 'Util/TimeUtil';
@@ -27,7 +28,6 @@ import {MessageHeader} from './ContentMessage/MessageHeader';
 import {MessageTime} from './MessageTime';
 
 import {DeleteMessage as DeleteMessageEntity} from '../../../entity/message/DeleteMessage';
-import {ServiceEntity} from '../../../integration/ServiceEntity';
 
 export interface DeleteMessageProps {
   message: DeleteMessageEntity;
