+++ src/script/components/ParticipantItemContent/ParticipantItemContent.tsx
@@ -24,8 +24,8 @@ import ko from 'knockout';
 import {UserBlockedBadge, UserVerificationBadges} from 'Components/Badge';
 import * as Icon from 'Components/Icon';
 import {UserInfo} from 'Components/UserInfo';
+import {ServiceEntity} from 'Repositories/integration/ServiceEntity';
 import {User} from 'src/script/entity/User';
-import {ServiceEntity} from 'src/script/integration/ServiceEntity';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
 
 import {
