+++ src/script/components/ServiceList/ServiceList.tsx
@@ -22,9 +22,8 @@ import React from 'react';
 import cx from 'classnames';
 
 import {ServiceListItem} from 'Components/ServiceList/components/ServiceListItem';
+import type {ServiceEntity} from 'Repositories/integration/ServiceEntity';
 import {t} from 'Util/LocalizerUtil';
-
-import type {ServiceEntity} from '../../integration/ServiceEntity';
 
 export enum MODE {
   COMPACT = 'ServiceList.MODE.COMPACT',
