+++ src/script/components/ServiceList/components/ServiceListItem/ServiceListItem.tsx
@@ -22,11 +22,10 @@ import {TabIndex} from '@wireapp/react-ui-kit/lib/types/enums';
 import {Avatar, AVATAR_SIZE} from 'Components/Avatar';
 import {ParticipantItemContent} from 'Components/ParticipantItemContent';
 import {listItem, listWrapper} from 'Components/ParticipantItemContent/ParticipantItem.styles';
+import {ServiceEntity} from 'Repositories/integration/ServiceEntity';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
 import {handleKeyDown} from 'Util/KeyboardUtil';
 import {t} from 'Util/LocalizerUtil';
-
-import {ServiceEntity} from '../../../../integration/ServiceEntity';
 
 export interface ServiceListItemProps {
   service: ServiceEntity;
