+++ src/script/components/panel/ServiceDetails.tsx
@@ -20,9 +20,8 @@
 import React from 'react';
 
 import {Avatar, AVATAR_SIZE} from 'Components/Avatar';
+import type {ServiceEntity} from 'Repositories/integration/ServiceEntity';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
-
-import type {ServiceEntity} from '../../integration/ServiceEntity';
 
 export interface ServiceDetailsProps {
   service: ServiceEntity;
