+++ src/script/guards/Panel.ts
@@ -17,8 +17,9 @@
  *
  */
 
+import {ServiceEntity} from 'Repositories/integration/ServiceEntity';
+
 import {User} from '../entity/User';
-import {ServiceEntity} from '../integration/ServiceEntity';
 import {PanelEntity} from '../page/RightSidebar/RightSidebar';
 
 export const isServiceEntity = (entity: PanelEntity): entity is ServiceEntity => {
