+++ src/script/main/app.ts
@@ -59,6 +59,8 @@ import {ServiceMiddleware} from 'Repositories/event/preprocessor/ServiceMiddlewa
 import {FederationEventProcessor} from 'Repositories/event/processor/FederationEventProcessor';
 import {GiphyRepository} from 'Repositories/giphy/GiphyRepository';
 import {GiphyService} from 'Repositories/giphy/GiphyService';
+import {IntegrationRepository} from 'Repositories/integration/IntegrationRepository';
+import {IntegrationService} from 'Repositories/integration/IntegrationService';
 import {initializeDataDog} from 'Util/DataDog';
 import {DebugUtil} from 'Util/DebugUtil';
 import {Environment} from 'Util/Environment';
@@ -81,8 +83,6 @@ import {BaseError} from '../error/BaseError';
 import {ClientError, CLIENT_ERROR_TYPE} from '../error/ClientError';
 import {TeamError} from '../error/TeamError';
 import {externalUrl} from '../externalRoute';
-import {IntegrationRepository} from '../integration/IntegrationRepository';
-import {IntegrationService} from '../integration/IntegrationService';
 import {startNewVersionPolling} from '../lifecycle/newVersionHandler';
 import {MediaRepository} from '../media/MediaRepository';
 import {initMLSGroupConversations, initialiseSelfAndTeamConversations} from '../mls';
