+++ src/script/page/LeftSidebar/panels/StartUI/ServicesTab.tsx
@@ -23,13 +23,13 @@ import {Button, ButtonVariant} from '@wireapp/react-ui-kit';
 
 import * as Icon from 'Components/Icon';
 import {ServiceList} from 'Components/ServiceList/ServiceList';
-import {IntegrationRepository} from 'src/script/integration/IntegrationRepository';
-import {ServiceEntity} from 'src/script/integration/ServiceEntity';
+import {useDebounce} from 'Hooks/useDebounce';
+import {IntegrationRepository} from 'Repositories/integration/IntegrationRepository';
+import {ServiceEntity} from 'Repositories/integration/ServiceEntity';
 import {t} from 'Util/LocalizerUtil';
 import {safeWindowOpen} from 'Util/SanitizationUtil';
 
 import {getManageServicesUrl} from '../../../../externalRoute';
-import {useDebounce} from '../../../../hooks/useDebounce';
 
 export const ServicesTab: React.FC<{
   canManageServices: boolean;
