+++ src/script/page/RightSidebar/ConversationDetails/ConversationDetails.test.tsx
@@ -24,6 +24,7 @@ import {ConnectionRepository} from 'Repositories/connection/ConnectionRepository
 import {ConversationRepository} from 'Repositories/conversation/ConversationRepository';
 import {ConversationRoleRepository} from 'Repositories/conversation/ConversationRoleRepository';
 import {MessageRepository} from 'Repositories/conversation/MessageRepository';
+import {IntegrationRepository} from 'Repositories/integration/IntegrationRepository';
 import {Conversation} from 'src/script/entity/Conversation';
 import {createUuid} from 'Util/uuid';
 
@@ -31,7 +32,6 @@ import {ConversationDetails} from './ConversationDetails';
 
 import {TestFactory} from '../../../../../test/helper/TestFactory';
 import {User} from '../../../entity/User';
-import {IntegrationRepository} from '../../../integration/IntegrationRepository';
 import {SearchRepository} from '../../../search/SearchRepository';
 import {SelfRepository} from '../../../self/SelfRepository';
 import {TeamEntity} from '../../../team/TeamEntity';
