+++ src/script/page/RightSidebar/ConversationDetails/components/ConversationDetailsHeader/ConversationDetailsHeader.test.tsx
@@ -20,8 +20,8 @@
 import {render, fireEvent} from '@testing-library/react';
 import {CONVERSATION_TYPE} from '@wireapp/api-client/lib/conversation';
 
+import {ServiceEntity} from 'Repositories/integration/ServiceEntity';
 import {User} from 'src/script/entity/User';
-import {ServiceEntity} from 'src/script/integration/ServiceEntity';
 import {createUuid} from 'Util/uuid';
 
 import {ConversationDetailsHeader} from './ConversationDetailsHeader';
