+++ src/script/page/RightSidebar/ConversationDetails/components/ConversationDetailsHeader/ConversationDetailsHeader.tsx
@@ -22,6 +22,7 @@ import {ChangeEvent, FC, KeyboardEvent, useEffect, useRef, useState} from 'react
 import {ConversationVerificationBadges} from 'Components/Badge';
 import * as Icon from 'Components/Icon';
 import {ConversationRepository} from 'Repositories/conversation/ConversationRepository';
+import {ServiceEntity} from 'Repositories/integration/ServiceEntity';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
 import {isEnterKey} from 'Util/KeyboardUtil';
 import {t} from 'Util/LocalizerUtil';
@@ -29,7 +30,6 @@ import {removeLineBreaks} from 'Util/StringUtil';
 
 import {Conversation} from '../../../../../entity/Conversation';
 import {User} from '../../../../../entity/User';
-import {ServiceEntity} from '../../../../../integration/ServiceEntity';
 import {GroupDetails} from '../GroupDetails/GroupDetails';
 
 interface ConversationDetailsHeaderProps {
