+++ src/script/page/RightSidebar/ConversationDetails/components/ConversationDetailsParticipants/ConversationDetailsParticipants.tsx
@@ -21,11 +21,11 @@ import * as Icon from 'Components/Icon';
 import {ServiceList} from 'Components/ServiceList';
 import {UserList} from 'Components/UserList';
 import {ConversationRepository} from 'Repositories/conversation/ConversationRepository';
+import {ServiceEntity} from 'Repositories/integration/ServiceEntity';
 import {t} from 'Util/LocalizerUtil';
 
 import {Conversation} from '../../../../../entity/Conversation';
 import {User} from '../../../../../entity/User';
-import {ServiceEntity} from '../../../../../integration/ServiceEntity';
 
 interface ConversationDetailsParticipantsProps {
   activeConversation: Conversation;
