+++ src/script/repositories/integration/ServiceEntity.ts
