+++ src/script/view_model/ActionsViewModel.ts
@@ -31,6 +31,8 @@ import type {ConnectionRepository} from 'Repositories/connection/ConnectionRepos
 import type {ConversationRepository} from 'Repositories/conversation/ConversationRepository';
 import type {MessageRepository} from 'Repositories/conversation/MessageRepository';
 import {NOTIFICATION_STATE} from 'Repositories/conversation/NotificationSetting';
+import type {IntegrationRepository} from 'Repositories/integration/IntegrationRepository';
+import type {ServiceEntity} from 'Repositories/integration/ServiceEntity';
 import {t} from 'Util/LocalizerUtil';
 import {isBackendError} from 'Util/TypePredicateUtil';
 
@@ -39,8 +41,6 @@ import type {MainViewModel} from './MainViewModel';
 import type {Conversation} from '../entity/Conversation';
 import type {Message} from '../entity/message/Message';
 import type {User} from '../entity/User';
-import type {IntegrationRepository} from '../integration/IntegrationRepository';
-import type {ServiceEntity} from '../integration/ServiceEntity';
 import {SelfRepository} from '../self/SelfRepository';
 import {UserState} from '../user/UserState';
