+++ src/script/view_model/MainViewModel.ts
@@ -30,13 +30,13 @@ import type {MessageRepository} from 'Repositories/conversation/MessageRepositor
 import type {CryptographyRepository} from 'Repositories/cryptography/CryptographyRepository';
 import type {EventRepository} from 'Repositories/event/EventRepository';
 import type {GiphyRepository} from 'Repositories/giphy/GiphyRepository';
+import type {IntegrationRepository} from 'Repositories/integration/IntegrationRepository';
 
 import {ActionsViewModel} from './ActionsViewModel';
 import {CallingViewModel} from './CallingViewModel';
 import {ContentViewModel} from './ContentViewModel';
 import {ListViewModel} from './ListViewModel';
 
-import type {IntegrationRepository} from '../integration/IntegrationRepository';
 import type {MediaRepository} from '../media/MediaRepository';
 import type {NotificationRepository} from '../notification/NotificationRepository';
 import type {PreferenceNotificationRepository} from '../notification/PreferenceNotificationRepository';
